Moved schemas to lib
