+++ src/lib/server/sessions/toastSession.ts
@@ -1,23 +1,15 @@
 import type { CookieSerializeOptions } from 'cookie'
 import { z } from 'zod'
-import { createId as cuid } from '@paralleldrive/cuid2'
 import type { Cookies } from '@sveltejs/kit'
 import {
 	decryptCookie,
 	encryptAndSignCookieValue,
 } from '$lib/server/sessions/secureCookie'
+import { ToastSchema, toastTypes } from '$lib/schemas'
 
-const types = ['message', 'success', 'error'] as const
-export type Type = (typeof types)[number]
+export type Type = (typeof toastTypes)[number]
 export type Toast = z.infer<typeof ToastSchema>
 export type ToastInput = z.input<typeof ToastSchema>
-
-export const ToastSchema = z.object({
-	description: z.string(),
-	id: z.string().default(() => cuid()),
-	title: z.string().optional(),
-	type: z.enum(types).default('message'),
-})
 
 export const toastCookieName = 'dn_toast'
 export const toastOptionValues: CookieSerializeOptions & {
