--- src/lib/utils/userValidation.ts
@@ -1,43 +1,0 @@
-import { z } from 'zod'
-
-export const UsernameSchema = z
-	.string({ required_error: 'Username is required' })
-	.min(3, { message: 'Username is too short' })
-	.max(20, { message: 'Username is too long' })
-	.regex(/^[a-zA-Z0-9_]+$/, {
-		message: 'Username can only include letters, numbers, and underscores',
-	})
-	.transform(value => value?.toLowerCase())
-
-export const PasswordSchema = z
-	.string({ required_error: 'Password is required' })
-	.min(6, { message: 'Password is too short' })
-	.max(100, { message: 'Password is too long' })
-
-export const NameSchema = z
-	.string({ required_error: 'Name is required' })
-	.min(3, { message: 'Name is too short' })
-	.max(40, { message: 'Name is too long' })
-
-export const PasswordAndConfirmPasswordSchema = z.object({
-	password: PasswordSchema,
-	confirm: PasswordSchema,
-})
-
-export const EmailSchema = z
-	.string({ required_error: 'Email is required' })
-	.email({ message: 'Email is invalid' })
-	.min(3, { message: 'Email is too short' })
-	.max(100, { message: 'Email is too long' })
-	// make sure to map email to lowercase
-	.transform(value => value.toLowerCase())
-
-export const UserNameOrEmailSchema = z
-	.string({ required_error: 'Username/Email is required' })
-	.min(3, { message: 'Username/Email is too short' })
-	.max(100, { message: 'Username/Email is too long' })
-	.regex(/^[a-zA-Z0-9_]+$/, {
-		message:
-			'Username/Email can only include letters, numbers, and underscores',
-	})
-	.transform(value => value.toLowerCase())
