+++ src/routes/users/[username]/notes/[noteId]/edit/+page.server.ts
@@ -2,7 +2,7 @@ import { superValidate } from 'sveltekit-superforms'
 import { zod } from 'sveltekit-superforms/adapters'
 import { type Actions } from '@sveltejs/kit'
 import { newOrUpdate } from '$lib/components/EditNote/newOrUpdate.server'
-import { NoteEditorSchema } from '$lib/components/EditNote/types'
+import { NoteEditorSchema } from '$lib/schemas'
 
 export const load = async ({ parent }) => {
 	const { note } = await parent()
