+++ src/routes/users/[username]/notes/new-note/+page.server.ts
@@ -2,8 +2,8 @@ import { superValidate } from 'sveltekit-superforms'
 import { zod } from 'sveltekit-superforms/adapters'
 import { type Actions } from '@sveltejs/kit'
 import { newOrUpdate } from '$lib/components/EditNote/newOrUpdate.server'
-import { NoteEditorSchema } from '$lib/components/EditNote/types'
 import type { PageServerLoad } from './$types'
+import { NoteEditorSchema } from '$lib/schemas'
 
 export const load = (async () => {
 	const newNoteForm = await superValidate(zod(NoteEditorSchema))
